+++ examples/user-guide/08-regridding.py
@@ -186,18 +186,12 @@ axes[0].set_ylabel("original")
 axes[1].set_ylabel("regridded")
 
 # %%
-# Set up the input needed for custom regridding including the method and the old grid.
+# Set up the input needed for custom regridding. Create a regridder
+# weight-cache. This object can (and should) be reused for all the packages that
+# undergo custom regridding at this stage.
 from imod.mf6.utilities.regrid import RegridderWeightsCache
 
-# just take any array of the package to use as the old grid
-old_grid = original_rch_package["rate"]
-
-old_grid
-
-# %%
-# Create a regridder weight-cache. This object can (and should) be reused for all the packages
-# that undergo custom regridding at this stage.
-regrid_context = RegridderWeightsCache(original_rch_package["rate"], target_grid)
+regrid_context = RegridderWeightsCache()
 
 regrid_context